backend/requirements.txt
fastapi
uvicorn
httpx
groq
python-dotenv
pydantic
pydantic[email]
motor
dnspython
passlib[bcrypt]==1.7.4
bcrypt<4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
